Modified the strctures used in LISAinference likelihood for better efficiency with log linear overlaps.
LISAinference, struct, likelihood: Numerous changes to optimize performance of loglinear overlaps. In the new implementation, the injection waveform, its frequencies and the noise values are kept as persistent structures. Doubled all structures, CAmpPhase and ReIm. The tag tagint decides to use the CAmpPhase or ReIm representation. Now accepts nbptsoverlap as a global parameter for the number of points to be used in loglinear overlaps.
Note: Functions FDListmodesOverlap, FDLogLikelihood are kept; they allow for the selection tagint=0 for the wip overlap, tagint=1 for the non-optimized log-linear integration, where every waveform is reevaluated (with hardcoded nbpts).