Added a new version of the wieghted inner product integration. Added options. Bugs corrected.
tools/fresnel.c: routines for the computation of overlap integrals, split in intervals, using quadratic spline approximation for the phase and either Taylor expansion in the phase or Fresnel integral.
tools/splinecoefficients.c: routines for the computation of splines in the form of explicit matrices of coefficients.
Corrected a number of bugs in the response (notably branching bug in l when computing the coefficients from modes to polarizations).
LISAinference: added parameters to rescale distance prior limits when asked to rescale the signal to achieve a particular SNR, and to use a flat distance prior instead of a flat prior in volume. Parameters of the run are now automatically written in a .out file for reference.