Added Fisher matrix computation with LISAinference_ptmcmc.  This has so-far been tested to see that it gives plausible results which are mostly consistent with variations in methodology.
The routine applies second order differencing with the differencing step set to be some small fraction (set by --Fisher_err_target) some a scalefactor (for each param).  The scalefactor starts with a generic scale which the ptmcmc prior provides, and then then the Fisher is recalculated with the scale based on the minus 1/2 power of the Fisher diagonal elements.  This is iterated until the result converges to the expected target level.

The code is not particularly fast.  The slowest part is getting the Distance derivs right (which we could do analytically, of course) with the case of using non-flatprior distance scaling.