Added the to compute the relevant band within some SNR tolerance for the model signal mode and to then limit the range of data included in the overlap calculation.  This makes something like a factor of 2 speed up in 22 only cases, and more with higher modes.
Also added code for decimating a LISADataFD set of data by a factor of
two.  This is tested im the current version of my test program and
found to give consistent results within reasonable limits.  For recent
tests the likelihood err grows linearly with the decimation factor,but
the speed improved significantly.

I also added code for estimating the relevant temporal extent for
the relevant range of some signal model.  This can be used with the
above eventually to enable a speed up with minimal cost.