More progress in the fast computation of likelihoods with data. I have worked on speeding up the calculation with minimal impact on the likelihood (tested in not-trival maximum likelihood config). Tolerating logL errors of about ~0.1, with SNR~800, I now get speeds: 2yr 1-mode: 10ms 10yr 1mode: 30ms 2yr 5mode: 60ms
As before, the general idea is that the model is realized as spline
data. Then within spline segments, the model can be quickly evaluated
on a uniform grid using a recursive trick for the phase part. That
is still not too great unless we can somehow reduce the number of
times we multiply-add to the data.  A tolerance is set of the size
of each phase-derivative term in phase spline expression, to allow
increasing df steps.  For best results, it is also necessary to
treat the amplitude terms in the splines carefully to give estimates
for *averaged* values over df, rather than central values. Since we
have already required small evolution in phase this averaging is
important for accuracy when the amplitude changes relatively fast
compared to phase.

These changes yield a speedup to ~120ms (2yr,5-modes) or 30ms for
1 mode.   To get faster we need a faster-than-linear way to sum
up the finely-sampled data.  Toward this, we can precompute
2^n times downsampled versions of the data.  Then with a little bit
of tricky bookkeeping we can replace the on-the-fly sum of consecutive
data with use of the pre-decimated data.  This is implemented two
ways, one which applies a combination of various scale data-chunks
to add up to the requested summed sequence (this has a fair bit of
overhead).  The second way first takes steps to align the required
downsampling with the grids. This is a bit faster.  These latter
tricks for the resumming yield numerically equivalent results.

The other trick (slightly aside these issues) is to minimize the
range of frequencies in the overlap integrals (especially for higher
modes) This is done by setting a target on the amount of summed snr^2
(estimated quickly from the splines).

The parameters controling the appromations are the snr^2 cut (hard-
coded in LISAutils.c) and the phase-step tolerance (hard-coded in
data_likelihood.c).  Current results are with each of these set to
0.001.